Add files via upload
changing the parser to use regex - not finished. 

for now it goes over all files and look for patterns that are pre-defined. it will take the patterns and change them (not yet done). 
the function will seperete the file into word/patterns. for now it puts it in an array list but in the future will use a dictionary to map the files for each term